E-WhiteBoard/Bakend-E-WhiteBoard/python_backend/requirements.txt
Flask
opencv-python
mediapipe
python-socketio
numpy
